requirements.txt
python-dotenv
textwrap3
python-escpos
openai
